+++ src/CharacterPanel/slice/state/ClassModifier.ts
@@ -1,4 +1,3 @@
-import { Category } from "./Category";
 import { CategoryCalculationType } from "./CategoryCalculationType";
 import { v4 as uuidv4 } from "uuid";
